Refactor message feature into components.
