+++ group-collaboration/src/components/MessageEditor/index.js
